fix: local과 test 프로필에서 작동이 안 되는 문제 해결
FCM 관련 빈 중 prod와 dev에서만 빈으로 할당해 local과 test에서는 빈을 불러오지 못하는 문제가 발생해, 이를 해결함
