+++ build.gradle
@@ -120,7 +120,7 @@ jacocoTestReport {
 							'**/*Application*',
 							'**/common/**',
 							'**/configuration/**',
-							'**/FCMNotificationService/**'
+							'**/*FCMNotificationService/**'
 					] + Qdomains)
 		}))
 	}
@@ -154,7 +154,7 @@ jacocoTestCoverageVerification {
 					'*.configuration.*',
 					'*.common.*',
 					'*.exception.*',
-					'*.FCMNotificationService.*'
+					'*.*FCMNotificationService.*'
 			] + Qdomains
 		}
 	}
