+++ src/test/resources/application.yml
@@ -1,4 +1,7 @@
 spring:
+  profiles:
+    default: test
+
   datasource:
     url: jdbc:h2:mem:testdb
     username: sa
